fix(protocol): harden framing and recovery validation
Context: Manual review found protocol and recovery issues that would block the specified strict validation path: FIPC frames used the wrong fixed header size, closed stream generations were forgotten, and FST replay did not enforce declared transaction counts or commit digests.

Implementation: Correct FIPC header sizing and CRC offsets, remember retired stream generations for legal reuse checks, remove warning-prone unused code, tighten byte material conversions and iterator offsets, check event stream open results, assert nodiscard setup calls in tests, and validate FST transaction count/digest during recovery.

Safety/invariants: Malformed FIPC frames now match the specified 40-byte header layout; stream IDs cannot be reused without a higher generation after close; recovery applies only committed transactions with matching count and digest; strict-warning risks are reduced.

Verification: git diff --check; unfinished-marker scan; targeted source scans for byte literals and unused cycle state; Python inventory reported 35 named tests, 43 cpp files, and 11 headers. Compiler-backed verification remains blocked by missing local CMake/C++ toolchain.

+++ src/state/state_codec.cpp
@@ -32,7 +32,7 @@ Result<StateRecord> StateCodec::decode(const Bytes& bytes, std::size_t* consumed
     return Status::error(ErrorCode::format, "state-codec", "truncated state record");
   }
   const auto len = read_u32_le(bytes, 0).value();
-  if (len < kStateHeaderSize || len > bytes.size()) {
+  if (static_cast<std::size_t>(len) < kStateHeaderSize || static_cast<std::size_t>(len) > bytes.size()) {
     return Status::error(ErrorCode::format, "state-codec", "invalid state record length");
   }
   Bytes header(bytes.begin(), bytes.begin() + static_cast<std::ptrdiff_t>(kStateHeaderSize));
@@ -47,7 +47,7 @@ Result<StateRecord> StateCodec::decode(const Bytes& bytes, std::size_t* consumed
   record.base_sequence = read_u64_le(bytes, 16).value();
   record.new_sequence = read_u64_le(bytes, 24).value();
   const auto payload_crc = read_u32_le(bytes, 36).value();
-  record.payload.assign(bytes.begin() + static_cast<std::ptrdiff_t>(kStateHeaderSize), bytes.begin() + len);
+  record.payload.assign(bytes.begin() + static_cast<std::ptrdiff_t>(kStateHeaderSize), bytes.begin() + static_cast<std::ptrdiff_t>(len));
   if (crc32c(record.payload) != payload_crc) {
     return Status::error(ErrorCode::format, "state-codec", "state payload crc mismatch");
   }
